requirements.txt
cryptocompare
requests
py-etherscan-api
flexpoolapi
influxdb
colorama